chore: renombrar migracion stub a merge_alembic_heads
El archivo se llamaba 'legacy_payments_stamp' porque originalmente se
creo asumiendo que HU-021 (payments) iba a entrar en paralelo. HU-021
quedo descartada, pero la migracion sigue siendo necesaria: main tiene
dos heads (a1b2c3d4e5f6 y b1e8a2d9f7c4) que esta migracion vacia
mergea para que HU-023 y HU-024 puedan apilar sus cambios encima.

Solo se renombra el archivo y se actualiza el docstring; el revision id
'a9b8c7d6e5f4' se conserva para no invalidar bases ya stampeadas.
